+++ lib/pio/open_flow.rb
@@ -16,8 +16,9 @@ module Pio
 
     def self.switch_version(version)
       [:Barrier, :Echo, :Features, :FlowMod, :Hello, :Match,
-       :PacketIn, :FlowRemoved, :PacketOut, :SendOutPort, :PortStatus, :Stats,
-       :FlowStats, :DescriptionStats, :AggregateStats, :Error].each do |each|
+       :PacketIn, :FlowRemoved, :PacketOut, :SendOutPort, :PortStatus,
+       :Stats, :FlowStats, :DescriptionStats, :AggregateStats,
+       :TableStats, :Error].each do |each|
         set_message_class_name each, version
         @version = version.to_s
       end