+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## develop (unreleased)
 ### New features
 * [#250](https://github.com/trema/pio/pull/250): Add Flow Removed message parser.
+* [#251](https://github.com/trema/pio/pull/251): Add Table Stats Request message generator.
 
 
 ## 0.27.2 (10/25/2015)